+++ src/spyglass/decoding/decoding_merge.py
@@ -84,83 +84,68 @@ class DecodingOutput(_Merge, SpyglassMixin):
                     except (PermissionError, FileNotFoundError):
                         logger.warning(f"Unable to remove {path}, skipping")
 
-    @property
-    def source_class_dict(self) -> dict:
-        """Dictionary of source class names to source classes
-
-        {
-            'ClusterlessDecodingV1': spy...ClusterlessDecodingV1,
-             'SortedSpikesDecodingV1': spy...SortedSpikesDecodingV1
-        }
-
-        Returns
-        -------
-        dict
-            Dictionary of source class names to source classes
-        """
-        if not self._source_class_dict:
-            self._ensure_dependencies_loaded()
-            module = inspect.getmodule(self)
-            for part_name in self.parts():
-                part_name = to_camel_case(part_name.split("__")[-1].strip("`"))
-                part = getattr(module, part_name)
-                self._source_class_dict[part_name] = part
-        return self._source_class_dict
+    @classmethod
+    def fetch_results(cls, key):
+        return cls().merge_get_parent_class(key).fetch_results()
 
     @classmethod
-    def _get_source_class(cls, key):
-        # CB: By making this a property, we can generate the source_class_dict
-        #     without a key. Previously failed on empty table
-        #     This demonstrates pipeline-specific implementation. See also
-        #     merge_restrict_class edits that centralize this logic.
-        return cls.source_class_dict[(cls & key).fetch1("source")]
+    def fetch_model(cls, key):
+        return cls().merge_get_parent_class(key).fetch_model()
 
     @classmethod
-    def load_results(cls, key):
+    def fetch_environments(cls, key):
         decoding_selection_key = cls.merge_get_parent(key).fetch1("KEY")
-        source_class = cls._get_source_class(key)
-        return (source_class & decoding_selection_key).load_results()
+        return (
+            cls()
+            .merge_get_parent_class(key)
+            .fetch_environments(decoding_selection_key)
+        )
 
     @classmethod
-    def load_model(cls, key):
+    def fetch_position_info(cls, key):
         decoding_selection_key = cls.merge_get_parent(key).fetch1("KEY")
-        source_class = cls._get_source_class(key)
-        return (source_class & decoding_selection_key).load_model()
+        return (
+            cls()
+            .merge_get_parent_class(key)
+            .fetch_position_info(decoding_selection_key)
+        )
 
     @classmethod
-    def load_environments(cls, key):
+    def fetch_linear_position_info(cls, key):
         decoding_selection_key = cls.merge_get_parent(key).fetch1("KEY")
-        source_class = cls._get_source_class(key)
-        return source_class.load_environments(decoding_selection_key)
+        return (
+            cls()
+            .merge_get_parent_class(key)
+            .fetch_linear_position_info(decoding_selection_key)
+        )
 
     @classmethod
-    def load_position_info(cls, key):
+    def fetch_spike_data(cls, key, filter_by_interval=True):
         decoding_selection_key = cls.merge_get_parent(key).fetch1("KEY")
-        source_class = cls._get_source_class(key)
-        return source_class.load_position_info(decoding_selection_key)
-
-    @classmethod
-    def load_linear_position_info(cls, key):
-        decoding_selection_key = cls.merge_get_parent(key).fetch1("KEY")
-        source_class = cls._get_source_class(key)
-        return source_class.load_linear_position_info(decoding_selection_key)
-
-    @classmethod
-    def load_spike_data(cls, key, filter_by_interval=True):
-        decoding_selection_key = cls.merge_get_parent(key).fetch1("KEY")
-        source_class = cls._get_source_class(key)
-        return source_class.load_linear_position_info(
-            decoding_selection_key, filter_by_interval=filter_by_interval
+        return (
+            cls()
+            .merge_get_parent_class(key)
+            .fetch_linear_position_info(
+                decoding_selection_key, filter_by_interval=filter_by_interval
+            )
         )
 
     @classmethod
     def create_decoding_view(cls, key, head_direction_name="head_orientation"):
-        results = cls.load_results(key)
-        posterior = results.acausal_posterior.unstack("state_bins").sum("state")
-        env = cls.load_environments(key)[0]
+        results = cls.fetch_results(key)
+        posterior = (
+            results.squeeze()
+            .acausal_posterior.unstack("state_bins")
+            .drop_sel(state=["Local", "No-Spike"], errors="ignore")
+            .sum("state")
+        )
+        posterior /= posterior.sum("position")
+        env = cls.fetch_environments(key)[0]
 
         if "x_position" in results.coords:
-            position_info, position_variable_names = cls.load_position_info(key)
+            position_info, position_variable_names = cls.fetch_position_info(
+                key
+            )
             # Not 1D
             bin_size = (
                 np.nanmedian(np.diff(np.unique(results.x_position.values))),
@@ -180,9 +165,8 @@ class DecodingOutput(_Merge, SpyglassMixin):
             (
                 position_info,
                 position_variable_names,
-            ) = cls.load_linear_position_info(key)
+            ) = cls.fetch_linear_position_info(key)
             return create_1D_decode_view(
                 posterior=posterior,
                 linear_position=position_info["linear_position"],
-                ref_time_sec=position_info.index[0],
             )