+++ src/spyglass/decoding/v0/clusterless.py
@@ -38,7 +38,6 @@ try:
     )
 except ImportError as e:
     logger.warning(e)
-from ripple_detection import get_multiunit_population_firing_rate
 from tqdm.auto import tqdm
 
 from spyglass.common.common_behav import (
@@ -55,12 +54,12 @@ from spyglass.decoding.v0.dj_decoder_conversion import (
     convert_classes_to_dict,
     restore_classes,
 )
-from spyglass.spikesorting.spikesorting_curation import (
+from spyglass.spikesorting.v0.spikesorting_curation import (
     CuratedSpikeSorting,
     CuratedSpikeSortingSelection,
     Curation,
 )
-from spyglass.spikesorting.spikesorting_sorting import (
+from spyglass.spikesorting.v0.spikesorting_sorting import (
     SpikeSorting,
     SpikeSortingSelection,
 )
@@ -622,88 +621,6 @@ class ClusterlessClassifierParameters(SpyglassMixin, dj.Manual):
     def fetch1(self, *args, **kwargs) -> dict:
         """Custom fetch1 to convert dicts to classes"""
         return restore_classes(super().fetch1(*args, **kwargs))
-
-
-@schema
-class MultiunitFiringRate(SpyglassMixin, dj.Computed):
-    """Computes the population multiunit firing rate from the spikes in
-    MarksIndicator."""
-
-    definition = """
-    -> UnitMarksIndicator
-    ---
-    -> AnalysisNwbfile
-    multiunit_firing_rate_object_id: varchar(40)
-    """
-
-    def make(self, key):
-        marks = (UnitMarksIndicator & key).fetch_xarray()
-        multiunit_spikes = (np.any(~np.isnan(marks.values), axis=1)).astype(
-            float
-        )
-        multiunit_firing_rate = pd.DataFrame(
-            get_multiunit_population_firing_rate(
-                multiunit_spikes, key["sampling_rate"]
-            ),
-            index=marks.time,
-            columns=["firing_rate"],
-        )
-
-        # Insert into analysis nwb file
-        nwb_analysis_file = AnalysisNwbfile()
-        key["analysis_file_name"] = nwb_analysis_file.create(
-            key["nwb_file_name"]
-        )
-
-        key["multiunit_firing_rate_object_id"] = (
-            nwb_analysis_file.add_nwb_object(
-                analysis_file_name=key["analysis_file_name"],
-                nwb_object=multiunit_firing_rate.reset_index(),
-            )
-        )
-
-        nwb_analysis_file.add(
-            nwb_file_name=key["nwb_file_name"],
-            analysis_file_name=key["analysis_file_name"],
-        )
-
-        self.insert1(key)
-
-    def fetch1_dataframe(self) -> pd.DataFrame:
-        """Convenience function for returning the first dataframe"""
-        return self.fetch_dataframe()[0]
-
-    def fetch_dataframe(self) -> list[pd.DataFrame]:
-        """Fetches the multiunit firing rate as a list of pandas dataframes"""
-        return [
-            data["multiunit_firing_rate"].set_index("time")
-            for data in self.fetch_nwb()
-        ]
-
-
-@schema
-class MultiunitHighSynchronyEventsParameters(SpyglassMixin, dj.Manual):
-    """Params to extract times of high mulitunit activity during immobility."""
-
-    definition = """
-    param_name : varchar(80) # a name for this set of parameters
-    ---
-    minimum_duration = 0.015 :  float # minimum duration of event (in seconds)
-    zscore_threshold = 2.0 : float    # threshold event must cross to be considered (in std. dev.)
-    close_event_threshold = 0.0 :  float # events closer than this will be excluded (in seconds)
-    """
-
-    def insert_default(self):
-        """Insert the default parameter set"""
-        self.insert1(
-            {
-                "param_name": "default",
-                "minimum_duration": 0.015,
-                "zscore_threshold": 2.0,
-                "close_event_threshold": 0.0,
-            },
-            skip_duplicates=True,
-        )
 
 
 def get_decoding_data_for_epoch(