+++ src/spyglass/decoding/v1/clusterless.py
@@ -107,14 +107,14 @@ class ClusterlessDecodingV1(SpyglassMixin, dj.Computed):
         (
             position_info,
             position_variable_names,
-        ) = self.load_position_info(key)
+        ) = self.fetch_position_info(key)
 
         # Get the waveform features for the selected units
         # Don't need to filter by interval since the non_local_detector code will do that
         (
             spike_times,
             spike_waveform_features,
-        ) = self.load_spike_data(key, filter_by_interval=False)
+        ) = self.fetch_spike_data(key, filter_by_interval=False)
 
         # Get the encoding and decoding intervals
         encoding_interval = (
@@ -274,14 +274,14 @@ class ClusterlessDecodingV1(SpyglassMixin, dj.Computed):
 
         DecodingOutput.insert1(orig_key, skip_duplicates=True)
 
-    def load_results(self):
+    def fetch_results(self):
         return ClusterlessDetector.load_results(self.fetch1("results_path"))
 
-    def load_model(self):
+    def fetch_model(self):
         return ClusterlessDetector.load_model(self.fetch1("classifier_path"))
 
     @staticmethod
-    def load_environments(key):
+    def fetch_environments(key):
         model_params = (
             DecodingParameters
             & {"decoding_param_name": key["decoding_param_name"]}
@@ -297,7 +297,7 @@ class ClusterlessDecodingV1(SpyglassMixin, dj.Computed):
         (
             position_info,
             position_variable_names,
-        ) = ClusterlessDecodingV1.load_position_info(key)
+        ) = ClusterlessDecodingV1.fetch_position_info(key)
         classifier = ClusterlessDetector(**decoding_params)
 
         classifier.initialize_environments(
@@ -337,7 +337,7 @@ class ClusterlessDecodingV1(SpyglassMixin, dj.Computed):
         )
 
     @staticmethod
-    def load_position_info(key):
+    def fetch_position_info(key):
         position_group_key = {
             "position_group_name": key["position_group_name"],
             "nwb_file_name": key["nwb_file_name"],
@@ -362,10 +362,10 @@ class ClusterlessDecodingV1(SpyglassMixin, dj.Computed):
         return position_info, position_variable_names
 
     @staticmethod
-    def load_linear_position_info(key):
-        environment = ClusterlessDecodingV1.load_environments(key)[0]
+    def fetch_linear_position_info(key):
+        environment = ClusterlessDecodingV1.fetch_environments(key)[0]
 
-        position_df = ClusterlessDecodingV1.load_position_info(key)[0]
+        position_df = ClusterlessDecodingV1.fetch_position_info(key)[0]
         position_variable_names = (PositionGroup & key).fetch1(
             "position_variables"
         )
@@ -390,7 +390,7 @@ class ClusterlessDecodingV1(SpyglassMixin, dj.Computed):
         )
 
     @staticmethod
-    def load_spike_data(key, filter_by_interval=True):
+    def fetch_spike_data(key, filter_by_interval=True):
         waveform_keys = (
             (
                 UnitWaveformFeaturesGroup.UnitFeatures
@@ -428,7 +428,7 @@ class ClusterlessDecodingV1(SpyglassMixin, dj.Computed):
     def get_spike_indicator(cls, key, time):
         time = np.asarray(time)
         min_time, max_time = time[[0, -1]]
-        spike_times = cls.load_spike_data(key)[0]
+        spike_times = cls.fetch_spike_data(key)[0]
         spike_indicator = np.zeros((len(time), len(spike_times)))
 
         for ind, times in enumerate(spike_times):
@@ -465,12 +465,12 @@ class ClusterlessDecodingV1(SpyglassMixin, dj.Computed):
         # TODO: allow specification of track graph
         # TODO: Handle decode intervals, store in table
 
-        classifier = self.load_model()
-        results = self.load_results()
+        classifier = self.fetch_model()
+        results = self.fetch_results()
         posterior = results.acausal_posterior.unstack("state_bins").sum("state")
 
         if getattr(classifier.environments[0], "track_graph") is not None:
-            linear_position_info = self.load_linear_position_info(
+            linear_position_info = self.fetch_linear_position_info(
                 self.fetch1("KEY")
             )
 
@@ -495,7 +495,7 @@ class ClusterlessDecodingV1(SpyglassMixin, dj.Computed):
                 classifier.environments[0].track_graph, *traj_data
             )
         else:
-            position_info = self.load_position_info(self.fetch1("KEY"))
+            position_info = self.fetch_position_info(self.fetch1("KEY"))
             map_position = analysis.maximum_a_posteriori_estimate(posterior)
 
             orientation_name = (