+++ src/spyglass/linearization/v1/main.py
@@ -173,7 +173,7 @@ class LinearizedPositionV1(SpyglassMixin, dj.Computed):
 
         self.insert1(key)
 
-        from ..merge import LinearizedPositionOutput
+        from spyglass.linearization.merge import LinearizedPositionOutput
 
         part_name = to_camel_case(self.table_name.split("__")[-1])