+++ src/spyglass/spikesorting/v1/__init__.py
@@ -1,22 +1,29 @@
-from ..imported import ImportedSpikeSorting
-from .artifact import (
+from spyglass.spikesorting.imported import ImportedSpikeSorting
+from spyglass.spikesorting.v1.artifact import (
     ArtifactDetection,
     ArtifactDetectionParameters,
     ArtifactDetectionSelection,
 )
-from .curation import CurationV1
-from .figurl_curation import FigURLCuration, FigURLCurationSelection
-from .metric_curation import (
+from spyglass.spikesorting.v1.curation import CurationV1
+from spyglass.spikesorting.v1.figurl_curation import (
+    FigURLCuration,
+    FigURLCurationSelection,
+)
+from spyglass.spikesorting.v1.metric_curation import (
     MetricCuration,
     MetricCurationParameters,
     MetricCurationSelection,
     MetricParameters,
     WaveformParameters,
 )
-from .recording import (
+from spyglass.spikesorting.v1.recording import (
     SortGroup,
     SpikeSortingPreprocessingParameters,
     SpikeSortingRecording,
     SpikeSortingRecordingSelection,
 )
-from .sorting import SpikeSorterParameters, SpikeSorting, SpikeSortingSelection
+from spyglass.spikesorting.v1.sorting import (
+    SpikeSorterParameters,
+    SpikeSorting,
+    SpikeSortingSelection,
+)